requirements.txt
numpy
pandas
rasterio
scikit-image
pillow
matplotlib
streamlit
opencv-python-headless
tqdm 